Eliminado zenstack

+++ src/app/(loggedin)/admin/adminActions.ts
@@ -2,31 +2,54 @@
 
 import { revalidatePath } from "next/cache";
 import { AdminQuery } from "./types";
-import { getCurrentProfilePrismaClient } from "@/lib/prisma_utils";
 import {ADMINS_PER_PAGE} from "@/lib/data/pagination";
+import prisma from "@/lib/prisma";
+import {AdministratorWithUser} from "@/lib/definitions/administrator";
+import {mapAdministratorWithUser} from "@/lib/data/mappings";
+import {Prisma} from "@prisma/client";
+import AdministratorWhereInput = Prisma.AdministratorWhereInput;
 
-
-export async function getAdmins({ page, dni, lastName }: AdminQuery) {
+export async function getAdmins({ page, dni, lastName }: AdminQuery): Promise<AdministratorWithUser[]> {
     try {
-        const prisma = await getCurrentProfilePrismaClient();
-        return await prisma.administrator.findMany({
+        const filters: AdministratorWhereInput[] = []
+
+        if(dni !== undefined) {
+            filters.push({
+                profile: {
+                    user: {
+                        dni: dni
+                    }
+                }
+            })
+        }
+        if(lastName !== undefined) {
+            filters.push({
+                profile: {
+                    user: {
+                        lastName: {
+                            contains: lastName,
+                            mode: 'insensitive'
+                        }
+                    }
+                }
+        })
+        }
+
+        const administrators = await prisma.administrator.findMany({
             skip: (page - 1) * ADMINS_PER_PAGE,
             take: ADMINS_PER_PAGE,
             where: {
-                ...(dni != undefined && { dni: dni }),
-                ...(lastName != undefined && {
-                    user: {
-                        lastName: {
-                          contains: lastName,
-                          mode: 'insensitive',
-                        },
-                    }
-                }),
+                AND: filters
             },
             include: {
-                user: true
+                profile: {
+                    include: {
+                        user: true
+                    }
+                }
             }
         });
+        return administrators.map(mapAdministratorWithUser);
     } catch (error) {
         console.error("Error fetching administrators:", error);
         return [];
@@ -35,7 +58,6 @@ export async function getAdmins({ page, dni, lastName }: AdminQuery) {
 
 export async function getTotalAdmins() {
     try {
-        const prisma = await getCurrentProfilePrismaClient();
         return await prisma.administrator.count();
     } catch (error) {
         console.error("Error fetching administrators:", error);
@@ -45,29 +67,43 @@ export async function getTotalAdmins() {
 
 export async function removeAdmin(id: number) {
     try {
-        const prisma = await getCurrentProfilePrismaClient();
         const administrator = await prisma.administrator.findUnique({
             where: {
                 id
+            },
+            include:{
+                profile: {
+                    include: {
+                        user: {
+                            include: {
+                                profiles: true
+                            }
+                        }
+                    }
+                }
             }
         });
+
+        if (!administrator) {
+            console.error("Error fetching administrator");
+            return false;
+        }
+
         await prisma.administrator.delete({
             where: {
                 id
             }
         });
-        const user = await prisma.user.findUnique({
+        await prisma.profile.delete({
             where: {
-                dni: administrator!.dni
-            },
-            include: {
-                profiles: true
+                id: administrator.profile.id
             }
-        });
-        if (user?.profiles.length === 0) {
+        })
+
+        if (administrator.profile.user.profiles.length === 1) {
             await prisma.user.delete({
                 where: {
-                    dni: administrator!.dni
+                    dni: administrator.profile.user.dni
                 }
             });
         }
@@ -79,41 +115,39 @@ export async function removeAdmin(id: number) {
     }
 }
 
-export async function getAdmin(id: number) {
-    const prisma = await getCurrentProfilePrismaClient()
-    return prisma.administrator.findFirst({
+export async function getAdmin(id: number): Promise<AdministratorWithUser | null> {
+    const admin = await prisma.administrator.findFirst({
         where: {
             id: id
         },
-        select: {
-            id: true,
-            dni: true,
-            phoneNumber: true,
-            address: true,
-            email: true,
-            user: {
-                select: {
-                    firstName: true,
-                    lastName: true
+        include: {
+            profile: {
+                include: {
+                    user: true
                 }
             }
         }
     });
+    return admin ? mapAdministratorWithUser(admin) : null;
 }
 
 export async function editAdmin(phoneNumber: string, address: string, email: string, name: string, lastname: string, id: number): Promise<boolean>  {
-    const prisma = await getCurrentProfilePrismaClient();
     try {
         return await prisma.$transaction(async (prisma) => {
             await prisma.administrator.update({
                 data: {
                     phoneNumber: phoneNumber,
-                    email: email,
                     address: address,
-                    user: {
+
+                    profile: {
                         update: {
-                            firstName: name,
-                            lastName: lastname
+                            email: email,
+                            user: {
+                                update: {
+                                    firstName: name,
+                                    lastName: lastname
+                                }
+                            }
                         }
                     }
                 },
